feat: improve zsh aliases
There was a comment on a youtube where someone had `alias cat=bat`
and the comment was complaining about renaming tools. For some reason I
agreed with it, and here we are.

I think its mainly just the idea of being used to the original names,
but that is not really a good enough reason to add that indirection.
I will reserve aliases for complex commands, or for adding default
options.